ml-projects/sentiment-analyzer/requirements.txt
flask>=2.0.0
nltk>=3.7
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
textblob>=0.17.0
plotly>=5.0.0
dash>=2.0.0
gunicorn>=20.0.0
python-dotenv>=0.19.0
tweepy>=4.0.0
requests>=2.25.0 